+++ spec/models/activity_spec.rb
@@ -326,6 +326,13 @@ RSpec.describe Activity, type: :model do
       end
     end
 
+    context "when objectives is blank" do
+      subject(:activity) { build(:programme_activity, objectives: nil) }
+      it "should not be valid" do
+        expect(activity.valid?(:objectives_step)).to be_falsey
+      end
+    end
+
     context "when sector category is blank" do
       subject(:activity) { build(:activity, sector_category: nil) }
       it "should not be valid" do